+++ server/config/toml.go
@@ -80,6 +80,9 @@ metrics-address = "{{ .JSONRPC.MetricsAddress }}"
 # Maximum number of bytes returned from eth_call or similar invocations.
 return-data-limit = {{ .JSONRPC.ReturnDataLimit }}
 
+# Upgrade height for fix of revert gas refund logic when transaction reverted.
+fix-revert-gas-refund-height = {{ .JSONRPC.FixRevertGasRefundHeight }}
+
 ###############################################################################
 ###                             TLS Configuration                           ###
 ###############################################################################